+++ angular-workspace/projects/ni/nimble-angular/src/directives/combobox/tests/nimble-combobox-control-value-accessor.directive.spec.ts
@@ -1,6 +1,6 @@
 import { Component, ElementRef, ViewChild } from '@angular/core';
 import { ComponentFixture, fakeAsync, TestBed, tick } from '@angular/core/testing';
-import { FormControl, FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { FormControl, FormGroup, FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { NimbleComboboxModule } from '../nimble-combobox.module';
 import { NimbleListOptionModule } from '../../list-option/nimble-list-option.module';
 import { waitTask } from '../../../async-test-utilities';
@@ -13,6 +13,13 @@ function setComboboxValue(combobox: Combobox, index: number): void {
     combobox.options[index].dispatchEvent(new Event('click', { bubbles: true }));
 }
 
+function updateComboboxWithText(combobox: Combobox, text: string): void {
+    combobox.control.value = text;
+    const inputEvent = new InputEvent('input', { data: text, inputType: 'insertText' });
+    combobox.inputHandler(inputEvent);
+    combobox.dispatchEvent(inputEvent);
+}
+
 interface TestModel {
     name: string;
     value: number;
@@ -22,7 +29,7 @@ describe('Nimble combobox control value accessor', () => {
     describe('when using option\'s [ngValue] binding on template-based form', () => {
         @Component({
             template: `
-                <nimble-combobox #combobox [(ngModel)]="selectedOption" (ngModelChange)="onModelValueChange($event)" [compareWith]="compareWith" [disabled]="selectDisabled">                    
+                <nimble-combobox #combobox [(ngModel)]="selectedOption" (ngModelChange)="onModelValueChange($event)" [compareWith]="compareWith" [disabled]="selectDisabled" autocomplete="both">                    
                     <nimble-list-option *ngFor="let option of selectOptions" [ngValue]="option">{{ option?.name ?? nullValueString }}</nimble-list-option>
                     <nimble-list-option [ngValue]="dynamicOption">{{ dynamicOption?.name }}</nimble-list-option>
                 </nimble-combobox>
@@ -184,9 +191,7 @@ describe('Nimble combobox control value accessor', () => {
             fixture.detectChanges();
             tick();
             processUpdates();
-            combobox.control.value = 'Dynamic Option 1';
-            combobox.control.dispatchEvent(new InputEvent('input', { data: 'Dynamic Option 1', inputType: 'insertText' }));
-            combobox.dispatchEvent(new KeyboardEvent('keydown', { key: 'Enter' }));
+            updateComboboxWithText(combobox, 'Dynamic Option 1');
             fixture.detectChanges();
             tick();
             processUpdates();
@@ -200,9 +205,7 @@ describe('Nimble combobox control value accessor', () => {
             tick();
             processUpdates();
 
-            combobox.control.value = 'Option 1';
-            combobox.control.dispatchEvent(new InputEvent('input', { data: 'Option 1', inputType: 'insertText' }));
-            combobox.dispatchEvent(new KeyboardEvent('keydown', { key: 'Enter' }));
+            updateComboboxWithText(combobox, 'Option 1');
             fixture.detectChanges();
             tick();
             processUpdates();
@@ -211,9 +214,7 @@ describe('Nimble combobox control value accessor', () => {
         }));
 
         it('and user enters aribtrary text for value, callback value is notFound', () => {
-            combobox.control.value = 'f';
-            combobox.control.dispatchEvent(new InputEvent('input', { data: 'f', inputType: 'insertText' }));
-            combobox.dispatchEvent(new KeyboardEvent('keydown', { key: 'Enter' }));
+            updateComboboxWithText(combobox, 'f');
             fixture.detectChanges();
 
             expect(testHostComponent.callbackValue).toEqual(OPTION_NOT_FOUND);
@@ -223,8 +224,8 @@ describe('Nimble combobox control value accessor', () => {
     describe('when using option\'s [ngValue] binding on Reactive form', () => {
         @Component({
             template: `
-                <form>
-                    <nimble-combobox #combobox [formControl]="selectedOption" (change)="onModelValueChange()" [compareWith]="compareWith" [disabled]="selectDisabled">                    
+                <form [formGroup]="form">
+                    <nimble-combobox #combobox [formControl]="selectedOption" [compareWith]="compareWith" [disabled]="selectDisabled" autocomplete="both">                    
                         <nimble-list-option *ngFor="let option of selectOptions" [ngValue]="option">{{ option?.name ?? nullValueString }}</nimble-list-option>
                         <nimble-list-option [ngValue]="dynamicOption">{{ dynamicOption?.name }}</nimble-list-option>
                     </nimble-combobox>
@@ -243,21 +244,19 @@ describe('Nimble combobox control value accessor', () => {
 
             public defaultOption = this.selectOptions[1];
             public selectedOption = new FormControl(this.defaultOption);
+            public form: FormGroup = new FormGroup({
+                selectOption: this.selectedOption
+            });
+
             public dynamicOption: TestModel = { name: 'Dynamic Option 1', value: 4 };
             public readonly nullValueString = 'null';
 
             public selectDisabled = false;
 
-            public callbackValue: unknown;
-
             public useDefaultOptions = true;
 
             public compareWith(option1: TestModel | null, option2: TestModel | null): boolean {
                 return (!!option1 && !!option2 && option1.value === option2.value) || (option1 === null && option2 === null);
-            }
-
-            public onModelValueChange(): void {
-                this.callbackValue = this.selectedOption.value;
             }
         }
 
@@ -385,30 +384,12 @@ describe('Nimble combobox control value accessor', () => {
             expect(testHostComponent.selectedOption.value).toBe(null);
         }));
 
-        it('model changes for option, text display of old option entered, callback value is notFound ', fakeAsync(() => {
+        it('model changes for option, text display of old option entered, selectedOption value is OPTION_NOT_FOUND ', fakeAsync(() => {
             testHostComponent.dynamicOption = { name: 'foo', value: 10 };
             fixture.detectChanges();
             tick();
             processUpdates();
-            combobox.control.value = 'Dynamic Option 1';
-            combobox.control.dispatchEvent(new InputEvent('input', { data: 'Dynamic Option 1', inputType: 'insertText' }));
-            combobox.dispatchEvent(new KeyboardEvent('keydown', { key: 'Enter' }));
-            fixture.detectChanges();
-            tick();
-            processUpdates();
-
-            expect(testHostComponent.callbackValue).toEqual(OPTION_NOT_FOUND);
-        }));
-
-        it('list-option is removed followed by text of removed option entered as value, then \'testHostComponent.selectedOption\' is set to notFound', fakeAsync(() => {
-            testHostComponent.selectOptions.splice(0, 1); // remove 'Option 1'
-            fixture.detectChanges();
-            tick();
-            processUpdates();
-
-            combobox.control.value = 'Option 1';
-            combobox.control.dispatchEvent(new InputEvent('input', { data: 'Option 1', inputType: 'insertText' }));
-            combobox.dispatchEvent(new KeyboardEvent('keydown', { key: 'Enter' }));
+            updateComboboxWithText(combobox, 'Dynamic Option 1');
             fixture.detectChanges();
             tick();
             processUpdates();
@@ -416,13 +397,30 @@ describe('Nimble combobox control value accessor', () => {
             expect(testHostComponent.selectedOption.value).toEqual(OPTION_NOT_FOUND);
         }));
 
-        it('and user enters aribtrary text for value, callback value is notFound', () => {
-            combobox.control.value = 'f';
-            combobox.control.dispatchEvent(new InputEvent('input', { data: 'f', inputType: 'insertText' }));
-            combobox.dispatchEvent(new KeyboardEvent('keydown', { key: 'Enter' }));
+        it('list-option is removed followed by text of removed option entered as value, selectedOption value is OPTION_NOT_FOUND', fakeAsync(() => {
+            testHostComponent.selectOptions.splice(0, 1); // remove 'Option 1'
+            fixture.detectChanges();
+            tick();
+            processUpdates();
+
+            updateComboboxWithText(combobox, 'Option 1');
+            fixture.detectChanges();
+            tick();
+            processUpdates();
+
+            expect(testHostComponent.selectedOption.value).toEqual(OPTION_NOT_FOUND);
+        }));
+
+        it('as user types, selectedOption value is updated', () => {
+            updateComboboxWithText(combobox, 'f');
             fixture.detectChanges();
 
-            expect(testHostComponent.callbackValue).toEqual(OPTION_NOT_FOUND);
+            expect(testHostComponent.selectedOption.value).toEqual(OPTION_NOT_FOUND);
+
+            updateComboboxWithText(combobox, 'O'); // value should autocomplete to 'Option 1'
+            fixture.detectChanges();
+
+            expect(testHostComponent.selectedOption.value).toEqual(testHostComponent.selectOptions[0]);
         });
     });
 });