+++ angular-workspace/projects/ni/nimble-angular/src/directives/dialog/nimble-dialog.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { NimbleDialogDirective } from './nimble-dialog.directive';
+
+import '@ni/nimble-components/dist/esm/dialog';
+
+@NgModule({
+    declarations: [NimbleDialogDirective],
+    imports: [
+        CommonModule
+    ],
+    exports: [NimbleDialogDirective]
+})
+export class NimbleDialogModule { }